Modify alx-posts.php to allow asc/dsc ordering
I think it would be helpful to allow ascending and descending ordering options. The default will still be dsc, so it shouldn't affect existing users or change the original intended behavior. For myself, it will be a nice way to create a "Did you miss this?" section in the right sidebar. The left will still show the most resent posts. I'm sure there are plenty of other cases where it could be useful.